redo name requests and bus exporting (breaking)
Remove `bus.export()` in favor of `bus.requestName()` and
`name.export()`. Breaking this up into two steps make it much more clear
what's actually going on since requesting a name is the async part of
the operation and the actual export is not. This will make working with
name request flags easier and make errors easier to handle because there
is a single entry point for each.

Breaking.