cache name owners and discriminate signals on owner
Create a name cache. Cache the name owner (the unique name) when the
user creates a proxy object client and keep the cache up to date when
the name owner changes on the NameOwnerChanged standard DBus signal.

Then when a signal comes in, make sure that the sender for the signal
actually owns the name based on the `sender` field in the message and
only emit the signal on the proxy interface if it matches.

This fixes a bug where if two external connections take different names
but export the same objects with the same interface names (common for
DBus interface standards), the signal would be emitted on all proxy
objects that emit for that path/interface combination regardless of
whether it was for the name they are listening on.

See the new test for a precise description of the error case.